created report for participant dropout report
I've completed the majority of the report, missing only the date of the missing visit. I just don't know where that variable would be found.

I've also had an issue with the imported Redcap data not displaying the correct information. For all the dropped participants that I imported, it is only displaying 2 as dropped and I can't understand why.